backend/requirements.txt
flask
easyocr
pillow
flask-cors